[merge] 15 feat implement memo features

+++ ngo/src/main/java/com/ngo/service/ScrapService.java
@@ -4,10 +4,11 @@ import com.ngo.common.ApiResponse;
 import com.ngo.common.exception.NotFoundException;
 import com.ngo.common.message.ErrorMessage;
 import com.ngo.common.message.SuccessMessage;
-import com.ngo.dto.ScrapDto;
-import com.ngo.dto.ScrapListDto;
+import com.ngo.dto.*;
+import com.ngo.model.Memo;
 import com.ngo.model.Scrap;
 import com.ngo.model.User;
+import com.ngo.repository.MemoRepository;
 import com.ngo.repository.ScrapRepository;
 import com.ngo.repository.UserRepository;
 import org.springframework.stereotype.Service;
@@ -19,11 +20,13 @@ public class ScrapService
 {
     private final ScrapRepository scrapRepository;
     private final UserRepository userRepository;
+    private final MemoRepository memoRepository;
 
-    public ScrapService(ScrapRepository scrapRepository, UserRepository userRepository)
+    public ScrapService(ScrapRepository scrapRepository, UserRepository userRepository, MemoRepository memoRepository)
     {
         this.scrapRepository = scrapRepository;
         this.userRepository = userRepository;
+        this.memoRepository = memoRepository;
     }
 
     /**
@@ -62,11 +65,14 @@ public class ScrapService
 
     public ApiResponse<ScrapDto> deleteScrap(Long userId, Long scrapId)
     {
-        User user = userRepository.findById(userId)
+        userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundException(ErrorMessage.USER_NOT_FOUND));
 
         Scrap scrap = scrapRepository.findById(scrapId)
                 .orElseThrow(() -> new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND));
+
+        if (!userId.equals(scrap.getUser().getUserId()))
+            throw new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND);
 
         scrapRepository.delete(scrap);
         return ApiResponse.success(SuccessMessage.DELETE_SCRAP_SUCCESS);
@@ -75,4 +81,61 @@ public class ScrapService
     /**
      * 메모 관리
      */
+
+    public ApiResponse<MemoListDto> getAllMemos(Long userId, Long scrapId)
+    {
+        userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.USER_NOT_FOUND));
+
+        Scrap scrap = scrapRepository.findById(scrapId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND));
+
+        if (!userId.equals(scrap.getUser().getUserId()))
+            throw new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND);
+
+        List<MemoGetDto> memoDtoList = memoRepository.findByScrap_ScrapId(scrapId).stream()
+                .map(MemoGetDto::build)
+                .toList();
+
+        return ApiResponse.success(SuccessMessage.GET_MEMO_SUCCESS, new MemoListDto(scrapId, memoDtoList));
+    }
+
+    public ApiResponse<MemoGetDto> postMemo(Long userId, Long scrapId, MemoDto memoDto)
+    {
+        userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.USER_NOT_FOUND));
+
+        Scrap scrap = scrapRepository.findById(scrapId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND));
+
+        if (!userId.equals(scrap.getUser().getUserId()))
+            throw new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND);
+
+        Memo memo = Memo.builder()
+                .content(memoDto.getContent())
+                .scrap(scrap)
+                .build();
+
+        memoRepository.save(memo);
+        return ApiResponse.success(SuccessMessage.POST_MEMO_SUCESS, MemoGetDto.build(memo));
+    }
+
+    public ApiResponse<MemoGetDto> deleteMemo(Long userId, Long scrapId, Long memoId)
+    {
+        userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.USER_NOT_FOUND));
+
+        Scrap scrap = scrapRepository.findById(scrapId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND));
+
+        if (!userId.equals(scrap.getUser().getUserId()))
+            throw new NotFoundException(ErrorMessage.SCRAP_NOT_FOUND);
+
+        Memo memo = memoRepository.findById(memoId)
+                .orElseThrow(() -> new NotFoundException(ErrorMessage.MEMO_NOT_FOUND));
+        memoRepository.delete(memo);
+
+        return ApiResponse.success(SuccessMessage.DELETE_MEMO_SUCCESS);
+    }
+
 }
